LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_4-expected.txt


FAIL Removing iframe from document removes it from history assert_array_equals: property 2, expected 1 but got 2
